[#98] Convert the FunctionCompositionTest from Xtend to Java.
